% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{climate_data}
\alias{climate_data}
\title{Retrieve climate data from GISS.}
\usage{
climate_data()
}
\value{
A dataframe of climate data
}
\description{
Retrieve climate data from GISS.
}
\examples{
data <- climate_data()

}
