% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_spiral.R
\name{climate_spiral}
\alias{climate_spiral}
\title{Create climate spiral plot}
\usage{
climate_spiral(data = NULL, viridis = FALSE, animate = TRUE)
}
\arguments{
\item{animate}{}
}
\description{
Create climate spiral plot
}
