% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_stripe.R
\name{climate_stripe}
\alias{climate_stripe}
\title{Plot climate stripes}
\usage{
climate_stripe(data = NULL)
}
\arguments{
\item{data}{Optional data input. If not provided, climate_stripe will use the NASA GISS data.}
}
\value{
A ggplot
}
\description{
Plot climate stripes
}
\examples{
gg <- climate_stripe()
}
