% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_tornado.R
\name{climate_tornado}
\alias{climate_tornado}
\title{Title}
\usage{
climate_tornado(data = NULL)
}
\arguments{
\item{data}{Optional data input. If not provided, climate_stripe will use the NASA GISS data.}
}
\value{
A ggplot object
}
\description{
Title
}
\examples{
gg <- climate_tornado()
}
