% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_spiral.R
\name{climate_spiral}
\alias{climate_spiral}
\title{Create climate spiral plot}
\usage{
climate_spiral(data = NULL, by_year = FALSE, animate = FALSE)
}
\arguments{
\item{data}{Optional data input. If not provided, climate_spiral will use the NASA GISS data.}

\item{by_year}{TRUE/FALSE - Colour the output by Year, or by temperature difference (DEFAULT)}

\item{animate}{TRUE/FALSE - Should the output be animated to an mp4}
}
\value{
A ggplot object
}
\description{
Create climate spiral plot
}
\examples{
gg <- climate_spiral()
}
